feat: create PRView and move contens from Requisition to PRView
